backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
certifi==2026.2.25
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
motor==3.7.1
packaging==26.1
paho-mqtt==2.1.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pymongo==4.16.0
pytest==9.0.3
python-dotenv==1.2.2
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
upstash-redis==1.7.0
uvicorn==0.46.0
